my-distributed-ml-project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
